refactor(common): move counter and changelog packages from pkg to common
- Move pkg/counter to common/counter
- Move pkg/changelog to common/changelog
- Update import paths in 8 files
- Update README.md dependency policy

These generic utilities have no Tracee-specific logic and fit the common module design principles.
