+++ common/counter/counter.go
