+++ pkg/events/ftrace.go
@@ -13,8 +13,8 @@ import (
 
 	"github.com/aquasecurity/tracee/common"
 	"github.com/aquasecurity/tracee/common/capabilities"
+	"github.com/aquasecurity/tracee/common/counter"
 	"github.com/aquasecurity/tracee/common/logger"
-	"github.com/aquasecurity/tracee/pkg/counter"
 	"github.com/aquasecurity/tracee/types/trace"
 )
